refactory: 폴더명 수정과 리뷰 반영
